refactor(spr,smach)*Moved* code from spr to smach states
